# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AlgDevs-AI Agent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, queued, running, paused, completed, failed, cancelled]
        - name: agentType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
  /tasks/{id}/cancel:
    post:
      operationId: cancelTask
      tags: [tasks]
      summary: Cancel a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/retry:
    post:
      operationId: retryTask
      tags: [tasks]
      summary: Retry a failed task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retried task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all workspaces
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Providers ─────────────────────────────────────────────────────────────
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List all AI providers with health status
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      tags: [providers]
      summary: Add a provider configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderInput"
      responses:
        "201":
          description: Created provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /providers/{id}:
    patch:
      operationId: updateProvider
      tags: [providers]
      summary: Update a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderUpdate"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
    delete:
      operationId: deleteProvider
      tags: [providers]
      summary: Delete a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /providers/{id}/test:
    post:
      operationId: testProvider
      tags: [providers]
      summary: Test a provider connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderTestResult"
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, agentType, status, priority, retryCount, maxRetries, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        agentType:
          type: string
          enum: [supervisor, planner, researcher, coder, reviewer]
        status:
          type: string
          enum: [pending, queued, running, paused, completed, failed, cancelled]
        priority:
          type: integer
        retryCount:
          type: integer
        maxRetries:
          type: integer
        output:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        sessionId:
          type: ["integer", "null"]
        planId:
          type: ["integer", "null"]
        workspaceId:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TaskInput:
      type: object
      required: [title, agentType]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        agentType:
          type: string
          enum: [supervisor, planner, researcher, coder, reviewer]
        priority:
          type: integer
        maxRetries:
          type: integer
        sessionId:
          type: ["integer", "null"]
        planId:
          type: ["integer", "null"]
        workspaceId:
          type: ["integer", "null"]
...
    Workspace:
      type: object
      required: [id, name, status, artifactCount, taskCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived, deleted]
        artifactCount:
          type: integer
        taskCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    WorkspaceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    # ── Provider schemas ─────────────────────────────────────────────────────
    Provider:
      type: object
      required: [id, name, providerType, isActive, isHealthy, priority, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        providerType:
          type: string
          enum: [gemini, openrouter, groq, openai, anthropic, custom]
        baseUrl:
          type: ["string", "null"]
        defaultModel:
          type: ["string", "null"]
        isActive:
          type: boolean
        isHealthy:
          type: boolean
        priority:
          type: integer
        latencyMs:
          type: ["integer", "null"]
        lastCheckedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ProviderInput:
      type: object
      required: [name, providerType]
      properties:
        name:
          type: string
          minLength: 1
        providerType:
          type: string
          enum: [gemini, openrouter, groq, openai, anthropic, custom]
        baseUrl:
          type: ["string", "null"]
        defaultModel:
          type: ["string", "null"]
        apiKey:
          type: ["string", "null"]
        priority:
          type: integer
    ProviderUpdate:
      type: object
      properties:
        name:
          type: string
        defaultModel:
          type: ["string", "null"]
        isActive:
          type: boolean
        priority:
          type: integer
        apiKey:
          type: ["string", "null"]
    ProviderTestResult:
      type: object
      required: [success, latencyMs]
      properties:
        success:
          type: boolean
        latencyMs:
          type: integer
        model:
          type: ["string", "null"]
        error:
          type: ["string", "null"]

    # ── Agent instance schemas ───────────────────────────────────────────────
...